requirements.txt
fastapi
uvicorn
openai
pdfplumber
paddleocr
pillow
numpy
opencv-python-headless
